[EI-127] [web] feat: Add MetadataListItem component to numerical guidance
